more window slam funnies

+++ code/game/objects/structures/window.dm
@@ -446,18 +446,38 @@
 			I.melee_attack_chain(user, src, params)
 		return TRUE
 
-	var/def_zone = ran_zone(BODY_ZONE_HEAD, 20)
+
+	var/def_zone = grab.target_zone
+	var/obj/item/bodypart/BP = affecting.get_bodypart(def_zone)
+	if(!BP)
+		return
 	var/blocked = affecting_mob.run_armor_check(def_zone, MELEE)
 	if(grab.current_grab.damage_stage < GRAB_NECK)
-		affecting_mob.visible_message(span_danger("<b>[user]</b> bashes <b>[affecting_mob]</b> against \the [src]!"))
-		if(prob(50 * ((100 - blocked/100))))
-			affecting_mob.Knockdown(4 SECONDS)
+		affecting_mob.visible_message(span_danger("<b>[user]</b> bashes <b>[affecting_mob]</b>'s [BP.plaintext_zone] against \the [src]!"))
+		switch(def_zone)
+			if(BODY_ZONE_HEAD, BODY_ZONE_CHEST)
+				if(prob(50 * ((100 - blocked/100))))
+					affecting_mob.Knockdown(4 SECONDS)
+			if(BODY_ZONE_L_ARM, BODY_ZONE_R_ARM)
+				if(prob(50 * ((100 - blocked/100))))
+					var/side = def_zone == BODY_ZONE_L_ARM ? LEFT_HANDS : RIGHT_HANDS
+					var/obj/item/I = affecting.get_held_items_for_side(side)
+					if(I)
+						affecting.dropItemToGround(I)
 		affecting_mob.apply_damage(20, BRUTE, def_zone, blocked)
 		take_damage(10)
 		qdel(grab)
 	else
-		affecting_mob.visible_message(span_danger("<b>[user]</b> crushes <b>[affecting_mob]</b> against \the [src]!"))
-		affecting_mob.Knockdown(10 SECONDS)
+		affecting_mob.visible_message(span_danger("<b>[user]</b> crushes <b>[affecting_mob]</b>'s [BP.plaintext_zone] against \the [src]!"))
+		switch(def_zone)
+			if(BODY_ZONE_HEAD, BODY_ZONE_CHEST)
+				affecting_mob.Knockdown(10 SECONDS)
+			if(BODY_ZONE_L_ARM, BODY_ZONE_R_ARM)
+				var/side = def_zone == BODY_ZONE_L_ARM ? LEFT_HANDS : RIGHT_HANDS
+				var/obj/item/I = affecting.get_held_items_for_side(side)
+				if(I)
+					affecting.dropItemToGround(I)
+
 		affecting_mob.apply_damage(20, BRUTE, def_zone, blocked)
 		take_damage(20)
 		qdel(grab)
